refactor(deck): move graph integration into existing arrow-layers module
