+++ modules/arrow-layers/src/index.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
 
+export {OrthographicView, type PickingInfo} from '@deck.gl/core';
 export {
   ArrowPolygonLayer,
   type ArrowPolygonColorInput,
@@ -19,3 +20,14 @@ export {
   type ArrowTextLayerProps
 } from './layers/arrow-text-layer';
 export type {ArrowLayerPickingInfo} from './layers/arrow-layer-types';
+export {LuGraphDeckEffect, type LuGraphDeckDataset} from './lugraph/lugraph-effect';
+export {
+  LUGRAPH_DECK_EDGE_SHADER,
+  LuGraphEdgeLayer,
+  type LuGraphEdgeLayerProps
+} from './lugraph/lugraph-edge-layer';
+export {
+  LUGRAPH_DECK_NODE_SHADER,
+  LuGraphNodeLayer,
+  type LuGraphNodeLayerProps
+} from './lugraph/lugraph-node-layer';
