+++ modules/deck-luspatial/src/index.ts
@@ -2,19 +2,7 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
 
-export {OrthographicView, type PickingInfo} from '@deck.gl/core';
 export {
   LuSpatialPointLayer,
   type LuSpatialPointLayerProps
 } from './layers/luspatial-point-layer';
-export {LuGraphDeckEffect, type LuGraphDeckDataset} from './lugraph/lugraph-effect';
-export {
-  LUGRAPH_DECK_EDGE_SHADER,
-  LuGraphEdgeLayer,
-  type LuGraphEdgeLayerProps
-} from './lugraph/lugraph-edge-layer';
-export {
-  LUGRAPH_DECK_NODE_SHADER,
-  LuGraphNodeLayer,
-  type LuGraphNodeLayerProps
-} from './lugraph/lugraph-node-layer';
