+++ modules/deck-luspatial/test/luspatial-point-layer.node.spec.ts
@@ -185,8 +185,7 @@ describe('@deck.gl-community/luspatial package boundary', () => {
       '@luma.gl/core',
       '@luma.gl/engine',
       '@luma.gl/experimental',
-      '@luma.gl/shadertools',
-      '@luma.gl/tables'
+      '@luma.gl/shadertools'
     ]);
   });
 });
